Writing all input data in the form of string inline with the xml read operation.
The values of the constants would be read as string from the xml and converted to double before use in the project. Changed the code to reflect the same by creating string variables storing numerical values. Also created loops to add symbols (names of variables and constants in an expression) for the expression parser and evaluator